Added classes for some intro tables and method to add some data to the DB for testing with later
Class models are in sql_models.py. Still having some issues with keeping these out of main so there's some redundant flask boiler plate at the top 
sql_admin_functions.py contains functions to populate the database with test data, I will be adding more to this later and possibly changing its name.
in main added some calls to get the DB going as well as a call to the random info function from admin_functions to populate the DB with 20 entries. Tested on current build as of 1-31 and does work.
However, nothing shows on the site yet have to investigate the database file with DB Browser (SQLlite) to see the entries. Will be working on that soon with probably a dummy html page.